Absolute Zero, a replacement spell for Glaciate.
This level 9 ice spell targets the enemy nearest the player (breaking
ties randomly and skipping firewood and tentacles) and instantly kills
it, turning it into an ice block. Because it always targets and kills
the closest monster, range and damage scaling with spellpower don't
work. Instead, noise decreases as spellpower increases, from an
ear-shattering 40 at zero power to 0 at maximum power (the spell also
always makes 4 noise at the player's location).

This is one of the glaciate replacement proposals that's been kicked
around on discord etc. Motivation for the change is that all four of the
level 9 direct damage spells are large aoe spells, and of the four,
glaciate seems to be the worst from both power level and interface
perspectives. Also, ice is supposed to be the least blasty school I
think? This spell is effectively single target damage, so there's some
overlap with lcs and friends, but it's somewhat different in function.

Some concerns with this spell: it's very feast-or-famine in terms of
effect. Really good against Cerebov, horrible against packs of any
significant size. If it's too good in the good situations, it could be
nerfed by lowering the range or by adding a second school (probably
hexes?); given the other effects available at levels 8 and 9 I don't
anticipate absolute zero being op. It's much harder to buff the spell if
it's too weak, and imo it's important that the new spell be at least as
good as glaciate, preferably a bit better. It's also possible that this
type of instant death effect will just lead to fundamentally bad
gameplay, but that's to be decided by playtesting. Finally, replacing
glaciate with absolute zero will nerf megazig characters. I don't
consider megazigs a 'real' part of the game but that's probably worth
noting.

Commiter's Note: A subsequent commit will adjust this spell to replace
the spellpower-scaled noise reduction with a scaling spell range from 1
to 7, with the spell typically having range 5 when it's first castable.
Depending on how this feels in practice, we may later change the spell
to have a fixed but smaller range of 5, possibly restoring the noise
reduction.

I wrapped commit comment lines and ran checkwhite within this commit.