Add range scaling for Absolute Zero
Replace the spellpower-scaled noise reduction with a spell range that
scales with spellpower. The range scales from 1 at power 0 to 7 with max
power, with the range typically being 5 when the spell is first
castable. This is a more natural way to scale the strength of the effect
with spellpower, and gives some reasonable incentive to train more Ice or
use spellpower enhancers.

Instead of scaling the noise, give the spell a casting noise of 9 and an
effect noise of 25, which is still quite loud.