Fix a spell range check for Absolute Zero
The tracer check for casting this spell didn't consider spellpower when
checking for a monster target for the spellcasting UI, hence it could
only see a monster if it was one square away. Remove this "tracer" usage
and rely on the usual range checks that properly consider spellpower.
This spell didn't actually need this tracer mode in the first place,
since the spellcasting UI code already handled LOS-range spells
properly.